workspace/paper-repos/arxiv_2605_06564/requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
torch>=2.0
d3rlpy>=2.3
pymc>=5.0
python-igraph>=0.11
networkx>=3.0
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65
joblib>=1.3
